chore: update

+++ packages/contracts/contracts/WagumiCats.sol
@@ -30,7 +30,8 @@ contract WagumiCats is ERC721, ReentrancyGuard, Ownable {
 
   /** MINTING **/
 
-  address public constant MULTI_SIG = 0xDCE4694e268bD83EA41B335320Ed11A684a1d7dB;
+  address public constant MULTI_SIG =
+    0xDCE4694e268bD83EA41B335320Ed11A684a1d7dB;
 
   uint256 public constant MAX_SUPPLY = 1000;
 
@@ -39,7 +40,7 @@ contract WagumiCats is ERC721, ReentrancyGuard, Ownable {
   function ownerBatchMint() public onlyOwner {
     uint256 _currentId = totalSupply();
 
-    for (uint i = 0; i < 10; i++) {
+    for (uint256 i = 0; i < 10; i++) {
       _safeMint(MULTI_SIG, (_currentId + i));
       supplyCounter.increment();
     }
